requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
yt-dlp>=2024.1.0
openai-whisper>=20231117
chromadb>=0.4.22
python-multipart>=0.0.6
pydantic>=2.5.3
groq>=0.4.0
openai>=1.0.0
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
